+++ src/main/java/cn/lambdalib2/CorePlugin.java
@@ -13,10 +13,7 @@ public class CorePlugin implements IFMLLoadingPlugin {
 
     @Override
     public String[] getASMTransformerClass() {
-        return new String[] {
-          //      "cn.lambdalib2.render.mc.PipelineTransformer",
-                "cn.lambdalib2.registry.impl.RegistryTransformer"
-        };
+        return new String[]{"cn.lambdalib2.registry.impl.RegistryTransformer"};
     }
 
     @Override